backend/foodgram/requirements.txt
Django==2.2.28
django-extra-fields==3.0.2
django-filter==21.1
djangorestframework==3.13.1
djoser==2.1.0
fpdf2==2.5.3
gunicorn==20.1.0
Pillow==9.1.0
psycopg2-binary==2.8.6
